Refactor CLI artifact and repair helpers
