+++ crates/perfgate-cli/src/main.rs
@@ -1,9 +1,15 @@
 //! perfgate CLI - entry point for all workflows.
 
+mod artifact_io;
+mod repair_context;
+
 use anyhow::Context;
+use artifact_io::{
+    atomic_write, load_optional_baseline_receipt, location_exists, read_json,
+    read_json_from_location, with_tokio_runtime, write_json, write_json_to_location,
+};
 use clap::{Args, Parser, Subcommand, ValueEnum};
 use glob::glob;
-use object_store::{ObjectStore, path::Path as ObjectPath};
 use perfgate::app as perfgate_app;
 use perfgate::domain as perfgate_domain;
 use perfgate::integrations::github::{self, CommentOptions, GitHubClient};
@@ -25,9 +31,8 @@ use perfgate_app::{
     PromoteRequest, PromoteUseCase, RatchetUseCase, ReportRequest, ReportUseCase, RunBenchRequest,
     RunBenchUseCase, ScenarioEvaluateInput, ScenarioEvaluateRequest, ScenarioUseCase,
     SensorReportBuilder, SystemClock, TradeoffEvaluateRequest, TradeoffUseCase, classify_error,
-    github_annotations, is_host_mismatch_reason, preview_lines, redact_command_for_diagnostics,
-    render_json_diff, render_markdown, render_markdown_template, render_terminal_diff,
-    render_tradeoff_markdown,
+    github_annotations, is_host_mismatch_reason, preview_lines, render_json_diff, render_markdown,
+    render_markdown_template, render_terminal_diff, render_tradeoff_markdown,
     watch::{Debouncer, WatchRunRequest, WatchState, execute_watch_run, render_watch_display},
 };
 use perfgate_client::types::auth::Role;
@@ -49,16 +54,17 @@ use perfgate_types::error::{AdapterError, ConfigValidationError, IoError, Perfga
 use perfgate_types::fingerprint::sha256_hex;
 use perfgate_types::{
     AggregateWeightMode, AggregationPolicy, BASELINE_REASON_NO_BASELINE, BaselineServerConfig,
-    ChangedFilesSummary, CompareReceipt, CompareRef, ConfigFile, DECISION_BUNDLE_SCHEMA_V1,
-    DECISION_INDEX_SCHEMA_V1, DecisionArtifactIndex, DecisionBundleArtifact,
-    DecisionBundleArtifactContent, DecisionBundleArtifactKind, DecisionBundleMetadata,
-    DecisionBundleReceipt, FailIfNOfM, HostMismatchPolicy, Metric, MetricStatus,
-    OtelSpanIdentifiers, PerfgateReport, ProbeCompareReceipt, ProbeReceipt,
-    REPAIR_CONTEXT_SCHEMA_V1, RatchetConfig, RepairContextReceipt, RepairGitMetadata,
-    RepairMetricBreach, RunReceipt, ScenarioConfigFile, ScenarioReceipt, SensorVerdictStatus,
-    ToolInfo, TradeoffReceipt, VerdictStatus,
+    CompareReceipt, CompareRef, ConfigFile, DECISION_BUNDLE_SCHEMA_V1, DECISION_INDEX_SCHEMA_V1,
+    DecisionArtifactIndex, DecisionBundleArtifact, DecisionBundleArtifactContent,
+    DecisionBundleArtifactKind, DecisionBundleMetadata, DecisionBundleReceipt, FailIfNOfM,
+    HostMismatchPolicy, Metric, MetricStatus, PerfgateReport, ProbeCompareReceipt, ProbeReceipt,
+    RatchetConfig, RunReceipt, ScenarioConfigFile, ScenarioReceipt, SensorVerdictStatus, ToolInfo,
+    TradeoffReceipt, VerdictStatus,
 };
 use regex::Regex;
+#[cfg(test)]
+use repair_context::parse_changed_files_summary;
+use repair_context::{git_metadata, maybe_write_repair_context, run_git_capture};
 use serde_json::Value as JsonValue;
 use std::collections::{BTreeMap, BTreeSet};
 use std::fs;
@@ -67,7 +73,6 @@ use std::process::Command as ProcessCommand;
 use std::process::ExitCode;
 use std::sync::Arc;
 use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
-use url::Url;
 
 const BASELINE_SERVER_NOT_CONFIGURED: &str = "baseline server is not configured; set `--baseline-server`, `PERFGATE_SERVER_URL`, or `[baseline_server].url` in `perfgate.toml`";
 const DEFAULT_FALLBACK_BASELINE_DIR: &str = "baselines";
@@ -7033,8 +7038,6 @@ fn open_browser(url: &str) {
 }
 
 fn execute_scale(args: ScaleArgs) -> anyhow::Result<()> {
-    use std::process::Command as ProcessCommand;
-
     let ScaleArgs {
         name,
         command,
@@ -10031,206 +10034,6 @@ fn write_check_artifacts(outcome: &CheckOutcome, pretty: bool) -> anyhow::Result
     Ok(())
 }
 
-fn maybe_write_repair_context(
-    outcome: &CheckOutcome,
-    baseline_path: Option<&Path>,
-    emit_requested: bool,
-    pretty: bool,
-) -> anyhow::Result<()> {
-    let should_emit = emit_requested
-        || matches!(
-            outcome.report.verdict.status,
-            VerdictStatus::Warn | VerdictStatus::Fail
-        );
-    if !should_emit {
-        return Ok(());
-    }
-
-    let repair = build_repair_context(outcome, baseline_path);
-    let out_path = outcome
-        .run_path
-        .parent()
-        .unwrap_or(Path::new(""))
-        .join("repair_context.json");
-    write_json(&out_path, &repair, pretty)?;
-    Ok(())
-}
-
-fn build_repair_context(
-    outcome: &CheckOutcome,
-    baseline_path: Option<&Path>,
-) -> RepairContextReceipt {
-    let breached_metrics = if let Some(compare) = &outcome.compare_receipt {
-        compare
-            .deltas
-            .iter()
-            .filter_map(|(metric, delta)| {
-                if !matches!(delta.status.as_str(), "warn" | "fail" | "skip") {
-                    return None;
-                }
-                let budget = compare.budgets.get(metric)?;
-                Some(RepairMetricBreach {
-                    metric: *metric,
-                    status: delta.status.as_str().to_string(),
-                    baseline: delta.baseline,
-                    current: delta.current,
-                    regression: delta.regression,
-                    fail_threshold: budget.threshold,
-                    warn_threshold: budget.warn_threshold,
-                })
-            })
-            .collect()
-    } else {
-        Vec::new()
-    };
-
-    let compare_path = outcome
-        .compare_path
-        .as_ref()
-        .map(|p| p.display().to_string());
-    let report_path = outcome.report_path.display().to_string();
-    let profile_path = outcome.report.profile_path.clone();
-    let otel_span = otel_span_from_env();
-    let git = git_metadata();
-    let changed_files = changed_files_summary();
-    let suggested = recommended_next_commands(outcome, baseline_path);
-
-    RepairContextReceipt {
-        schema: REPAIR_CONTEXT_SCHEMA_V1.to_string(),
-        benchmark: outcome.run_receipt.bench.name.clone(),
-        verdict: outcome.report.verdict.clone(),
-        status: outcome.report.verdict.status,
-        breached_metrics,
-        compare_receipt_path: compare_path,
-        report_path,
-        profile_path,
-        git,
-        changed_files,
-        otel_span,
-        recommended_next_commands: suggested,
-    }
-}
-
-fn recommended_next_commands(outcome: &CheckOutcome, baseline_path: Option<&Path>) -> Vec<String> {
-    let mut cmds = Vec::new();
-    let rerun_cmd = redact_command_for_diagnostics(&outcome.run_receipt.bench.command).join(" ");
-    if !rerun_cmd.is_empty() {
-        cmds.push(format!("rerun current command: {rerun_cmd}"));
-    }
-    if let Some(compare_path) = &outcome.compare_path {
-        cmds.push(format!(
-            "perfgate explain --compare {}",
-            compare_path.display()
-        ));
-    }
-    cmds.push(format!(
-        "perfgate paired --name {} --baseline-cmd \"<baseline-cmd>\" --current-cmd \"<current-cmd>\" --repeat {} --out {}/paired.json",
-        outcome.run_receipt.bench.name,
-        outcome.run_receipt.bench.repeat.max(10),
-        outcome.run_path.parent().unwrap_or(Path::new("")).display()
-    ));
-    if let Some(base) = baseline_path {
-        cmds.push(format!(
-            "perfgate compare --baseline {} --current {} --out {}/recompare.json",
-            base.display(),
-            outcome.run_path.display(),
-            outcome.run_path.parent().unwrap_or(Path::new("")).display()
-        ));
-    }
-    cmds.push(
-        "perfgate bisect --good <good-ref> --bad HEAD --executable <bench-binary>".to_string(),
-    );
-    cmds
-}
-
-fn otel_span_from_env() -> Option<OtelSpanIdentifiers> {
-    let trace_id = std::env::var("OTEL_TRACE_ID").ok();
-    let span_id = std::env::var("OTEL_SPAN_ID").ok();
-    if trace_id.is_none() && span_id.is_none() {
-        None
-    } else {
-        Some(OtelSpanIdentifiers { trace_id, span_id })
-    }
-}
-
-fn git_metadata() -> Option<RepairGitMetadata> {
-    let branch = run_git_capture(&["rev-parse", "--abbrev-ref", "HEAD"]);
-    let sha = run_git_capture(&["rev-parse", "HEAD"]);
-    if branch.is_none() && sha.is_none() {
-        None
-    } else {
-        Some(RepairGitMetadata { branch, sha })
-    }
-}
-
-fn changed_files_summary() -> Option<ChangedFilesSummary> {
-    let output = run_git_capture_bytes(&["status", "--porcelain", "-z"])?;
-    Some(parse_changed_files_summary(&output))
-}
-
-fn parse_changed_files_summary(output: &[u8]) -> ChangedFilesSummary {
-    let mut files = Vec::new();
-    let mut by_top = BTreeMap::new();
-
-    let mut entries = output
-        .split(|byte| *byte == b'\0')
-        .filter(|entry| !entry.is_empty());
-    while let Some(entry) = entries.next() {
-        if entry.len() <= 3 {
-            continue;
-        }
-
-        let status = &entry[..2];
-        let current_path = if status.iter().any(|code| matches!(code, b'R' | b'C')) {
-            entries.next().unwrap_or(&[])
-        } else {
-            &entry[3..]
-        };
-
-        if current_path.is_empty() {
-            continue;
-        }
-
-        let path = String::from_utf8_lossy(current_path).into_owned();
-        files.push(path.clone());
-        let top = path
-            .split(['/', '\\'])
-            .next()
-            .filter(|s| !s.is_empty())
-            .unwrap_or(".")
-            .to_string();
-        *by_top.entry(top).or_insert(0) += 1;
-    }
-
-    ChangedFilesSummary {
-        file_count: files.len() as u32,
-        files,
-        file_count_by_top_level: by_top,
-    }
-}
-
-fn run_git_capture(args: &[&str]) -> Option<String> {
-    let output = ProcessCommand::new("git").args(args).output().ok()?;
-    if !output.status.success() {
-        return None;
-    }
-    let text = String::from_utf8(output.stdout).ok()?;
-    let trimmed = text.trim().to_string();
-    if trimmed.is_empty() {
-        None
-    } else {
-        Some(trimmed)
-    }
-}
-
-fn run_git_capture_bytes(args: &[&str]) -> Option<Vec<u8>> {
-    let output = ProcessCommand::new("git").args(args).output().ok()?;
-    if !output.status.success() {
-        return None;
-    }
-    Some(output.stdout)
-}
-
 fn execute_export(
     run: Option<PathBuf>,
     compare: Option<PathBuf>,
@@ -10485,162 +10288,6 @@ fn normalize_paired_cli_command(args: Vec<String>, flag_name: &str) -> anyhow::R
     }
 
     Ok(args)
-}
-
-struct RemoteLocation {
-    store: Arc<dyn ObjectStore>,
-    object_path: ObjectPath,
-}
-
-fn parse_remote_location(path: &Path) -> anyhow::Result<Option<RemoteLocation>> {
-    let uri = path.to_string_lossy().to_string();
-    if !is_remote_storage_uri(&uri) {
-        return Ok(None);
-    }
-
-    let url = Url::parse(&uri).with_context(|| format!("invalid remote URI {}", uri))?;
-    let (store, object_path) =
-        object_store::parse_url(&url).with_context(|| format!("parse remote URI {}", uri))?;
-
-    Ok(Some(RemoteLocation {
-        store: store.into(),
-        object_path,
-    }))
-}
-
-fn with_tokio_runtime<T, F>(f: F) -> anyhow::Result<T>
-where
-    F: std::future::Future<Output = anyhow::Result<T>>,
-{
-    let rt = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .context("initialize async runtime")?;
-    rt.block_on(f)
-}
-
-fn is_object_not_found(err: &object_store::Error) -> bool {
-    matches!(err, object_store::Error::NotFound { .. })
-        || err.to_string().to_ascii_lowercase().contains("not found")
-}
-
-fn location_exists(path: &Path) -> anyhow::Result<bool> {
-    if let Some(remote) = parse_remote_location(path)? {
-        let head = with_tokio_runtime(async move {
-            remote
-                .store
-                .head(&remote.object_path)
-                .await
-                .map_err(anyhow::Error::from)
-        });
-        return match head {
-            Ok(_) => Ok(true),
-            Err(err) => {
-                if err
-                    .downcast_ref::<object_store::Error>()
-                    .is_some_and(is_object_not_found)
-                {
-                    Ok(false)
-                } else {
-                    Err(err).with_context(|| format!("check existence {}", path.display()))
-                }
-            }
-        };
-    }
-    Ok(path.exists())
-}
-
-fn read_json_from_location<T: serde::de::DeserializeOwned>(path: &Path) -> anyhow::Result<T> {
-    if let Some(remote) = parse_remote_location(path)? {
-        let bytes = with_tokio_runtime(async move {
-            let result = remote
-                .store
-                .get(&remote.object_path)
-                .await
-                .map_err(anyhow::Error::from)?;
-            result.bytes().await.map_err(anyhow::Error::from)
-        })
-        .with_context(|| format!("read {}", path.display()))?;
-
-        return serde_json::from_slice(&bytes)
-            .with_context(|| format!("parse json {}", path.display()));
-    }
-
-    read_json(path)
-}
-
-fn write_json_to_location<T: serde::Serialize>(
-    path: &Path,
-    value: &T,
-    pretty: bool,
-) -> anyhow::Result<()> {
-    if let Some(remote) = parse_remote_location(path)? {
-        let bytes = if pretty {
-            serde_json::to_vec_pretty(value)?
-        } else {
-            serde_json::to_vec(value)?
-        };
-
-        with_tokio_runtime(async move {
-            remote
-                .store
-                .put(&remote.object_path, bytes.into())
-                .await
-                .map(|_| ())
-                .map_err(anyhow::Error::from)
-        })
-        .with_context(|| format!("write {}", path.display()))?;
-        return Ok(());
-    }
-
-    write_json(path, value, pretty)
-}
-
-fn load_optional_baseline_receipt(path: &Path) -> anyhow::Result<Option<RunReceipt>> {
-    if location_exists(path)? {
-        Ok(Some(read_json_from_location(path)?))
-    } else {
-        Ok(None)
-    }
-}
-
-fn read_json<T: serde::de::DeserializeOwned>(path: &Path) -> anyhow::Result<T> {
-    Ok(perfgate_types::read_json_file(path)?)
-}
-
-fn write_json<T: serde::Serialize>(path: &Path, value: &T, pretty: bool) -> anyhow::Result<()> {
-    let parent = path.parent().unwrap_or_else(|| Path::new(""));
-    if !parent.as_os_str().is_empty() {
-        fs::create_dir_all(parent).with_context(|| format!("create dir {}", parent.display()))?;
-    }
-
-    let bytes = if pretty {
-        serde_json::to_vec_pretty(value)?
-    } else {
-        serde_json::to_vec(value)?
-    };
-
-    atomic_write(path, &bytes)
-}
-
-fn atomic_write(path: &Path, bytes: &[u8]) -> anyhow::Result<()> {
-    use std::io::Write;
-
-    let parent = path.parent().unwrap_or_else(|| Path::new("."));
-    let mut tmp = parent.to_path_buf();
-    tmp.push(format!(".{}.tmp", uuid::Uuid::new_v4()));
-
-    {
-        let mut f =
-            fs::File::create(&tmp).with_context(|| format!("create temp {}", tmp.display()))?;
-        f.write_all(bytes)
-            .with_context(|| format!("write temp {}", tmp.display()))?;
-        f.sync_all().ok();
-    }
-
-    fs::rename(&tmp, path)
-        .with_context(|| format!("rename {} -> {}", tmp.display(), path.display()))?;
-    Ok(())
 }
 
 #[cfg(test)]
